Tests/LibWeb/Text/expected/XHR/XMLHttpRequest-null-body-after-non-null.txt

POST 200
GET 200
